Camera & Movement v2 (working)
Updated camera and movement system to make use of CineMachine. Replaced what was the previous "Main Camera" with a CM Virtual cam amd carried the PlayerCam script over to the new virtual camera. Now the Virtual cam serves as the main unless another virtual of higher priority is added to the scene.